continue adapting tanstack query to plan details
