+++ components/cards/IntroCard.tsx
